requirements.txt
Flask==1.1.2
joblib==1.2.0
librosa
pandas==1.5.3
python-dotenv==1.0.0
requests==2.24.0
xgboost==1.7.2
gunicorn
